Split mail sender sub package from mailer service package
