+++ services/mailer/sender/message_test.go
@@ -1,7 +1,7 @@
-// Copyright 2021 The Gogs Authors. All rights reserved.
+// Copyright 2024 The Gitea Authors. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-package mailer
+package sender
 
 import (
 	"strings"
